chore(storybook): extract common webpack config
